Revert "Re-generate wayland protocols"
This reverts commit c4d56751417961748d6ceb75b82e2f5fe06c8f50.
