+++ v3.3/glfw/glfw/src/wayland-idle-inhibit-unstable-v1-client-protocol.h
@@ -1,4 +1,4 @@
-/* Generated by wayland-scanner 1.19.0 */
+/* Generated by wayland-scanner */
 
 #ifndef IDLE_INHIBIT_UNSTABLE_V1_CLIENT_PROTOCOL_H
 #define IDLE_INHIBIT_UNSTABLE_V1_CLIENT_PROTOCOL_H
@@ -45,8 +45,6 @@ struct wl_surface;
 struct zwp_idle_inhibit_manager_v1;
 struct zwp_idle_inhibitor_v1;
 
-#ifndef ZWP_IDLE_INHIBIT_MANAGER_V1_INTERFACE
-#define ZWP_IDLE_INHIBIT_MANAGER_V1_INTERFACE
 /**
  * @page page_iface_zwp_idle_inhibit_manager_v1 zwp_idle_inhibit_manager_v1
  * @section page_iface_zwp_idle_inhibit_manager_v1_desc Description
@@ -83,9 +81,6 @@ struct zwp_idle_inhibitor_v1;
  * interface version number is reset.
  */
 extern const struct wl_interface zwp_idle_inhibit_manager_v1_interface;
-#endif
-#ifndef ZWP_IDLE_INHIBITOR_V1_INTERFACE
-#define ZWP_IDLE_INHIBITOR_V1_INTERFACE
 /**
  * @page page_iface_zwp_idle_inhibitor_v1 zwp_idle_inhibitor_v1
  * @section page_iface_zwp_idle_inhibitor_v1_desc Description
@@ -122,7 +117,6 @@ extern const struct wl_interface zwp_idle_inhibit_manager_v1_interface;
  * de-idles and re-idles the inhibitor will take effect.
  */
 extern const struct wl_interface zwp_idle_inhibitor_v1_interface;
-#endif
 
 #define ZWP_IDLE_INHIBIT_MANAGER_V1_DESTROY 0
 #define ZWP_IDLE_INHIBIT_MANAGER_V1_CREATE_INHIBITOR 1
