+++ v3.3/glfw/glfw/src/wayland-relative-pointer-unstable-v1-client-protocol.h
@@ -1,4 +1,4 @@
-/* Generated by wayland-scanner 1.19.0 */
+/* Generated by wayland-scanner */
 
 #ifndef RELATIVE_POINTER_UNSTABLE_V1_CLIENT_PROTOCOL_H
 #define RELATIVE_POINTER_UNSTABLE_V1_CLIENT_PROTOCOL_H
@@ -73,8 +73,6 @@ struct wl_pointer;
 struct zwp_relative_pointer_manager_v1;
 struct zwp_relative_pointer_v1;
 
-#ifndef ZWP_RELATIVE_POINTER_MANAGER_V1_INTERFACE
-#define ZWP_RELATIVE_POINTER_MANAGER_V1_INTERFACE
 /**
  * @page page_iface_zwp_relative_pointer_manager_v1 zwp_relative_pointer_manager_v1
  * @section page_iface_zwp_relative_pointer_manager_v1_desc Description
@@ -91,9 +89,6 @@ struct zwp_relative_pointer_v1;
  * given pointer.
  */
 extern const struct wl_interface zwp_relative_pointer_manager_v1_interface;
-#endif
-#ifndef ZWP_RELATIVE_POINTER_V1_INTERFACE
-#define ZWP_RELATIVE_POINTER_V1_INTERFACE
 /**
  * @page page_iface_zwp_relative_pointer_v1 zwp_relative_pointer_v1
  * @section page_iface_zwp_relative_pointer_v1_desc Description
@@ -114,7 +109,6 @@ extern const struct wl_interface zwp_relative_pointer_manager_v1_interface;
  * focus.
  */
 extern const struct wl_interface zwp_relative_pointer_v1_interface;
-#endif
 
 #define ZWP_RELATIVE_POINTER_MANAGER_V1_DESTROY 0
 #define ZWP_RELATIVE_POINTER_MANAGER_V1_GET_RELATIVE_POINTER 1
