+++ v3.3/glfw/glfw/src/wayland-viewporter-client-protocol.c
@@ -1,4 +1,4 @@
-/* Generated by wayland-scanner 1.19.0 */
+/* Generated by wayland-scanner */
 
 /*
  * Copyright © 2013-2016 Collabora, Ltd.
