+++ v3.3/glfw/glfw/src/wayland-viewporter-client-protocol.h
@@ -1,4 +1,4 @@
-/* Generated by wayland-scanner 1.19.0 */
+/* Generated by wayland-scanner */
 
 #ifndef VIEWPORTER_CLIENT_PROTOCOL_H
 #define VIEWPORTER_CLIENT_PROTOCOL_H
@@ -45,8 +45,6 @@ struct wl_surface;
 struct wp_viewport;
 struct wp_viewporter;
 
-#ifndef WP_VIEWPORTER_INTERFACE
-#define WP_VIEWPORTER_INTERFACE
 /**
  * @page page_iface_wp_viewporter wp_viewporter
  * @section page_iface_wp_viewporter_desc Description
@@ -71,9 +69,6 @@ struct wp_viewporter;
  * surface size.
  */
 extern const struct wl_interface wp_viewporter_interface;
-#endif
-#ifndef WP_VIEWPORT_INTERFACE
-#define WP_VIEWPORT_INTERFACE
 /**
  * @page page_iface_wp_viewport wp_viewport
  * @section page_iface_wp_viewport_desc Description
@@ -208,7 +203,6 @@ extern const struct wl_interface wp_viewporter_interface;
  * on the next wl_surface.commit.
  */
 extern const struct wl_interface wp_viewport_interface;
-#endif
 
 #ifndef WP_VIEWPORTER_ERROR_ENUM
 #define WP_VIEWPORTER_ERROR_ENUM
