+++ v3.3/glfw/glfw/src/wayland-xdg-shell-client-protocol.h
@@ -1,4 +1,4 @@
-/* Generated by wayland-scanner 1.19.0 */
+/* Generated by wayland-scanner */
 
 #ifndef XDG_SHELL_CLIENT_PROTOCOL_H
 #define XDG_SHELL_CLIENT_PROTOCOL_H
@@ -58,8 +58,6 @@ struct xdg_surface;
 struct xdg_toplevel;
 struct xdg_wm_base;
 
-#ifndef XDG_WM_BASE_INTERFACE
-#define XDG_WM_BASE_INTERFACE
 /**
  * @page page_iface_xdg_wm_base xdg_wm_base
  * @section page_iface_xdg_wm_base_desc Description
@@ -82,9 +80,6 @@ struct xdg_wm_base;
  * creating transient windows such as popup menus.
  */
 extern const struct wl_interface xdg_wm_base_interface;
-#endif
-#ifndef XDG_POSITIONER_INTERFACE
-#define XDG_POSITIONER_INTERFACE
 /**
  * @page page_iface_xdg_positioner xdg_positioner
  * @section page_iface_xdg_positioner_desc Description
@@ -135,9 +130,6 @@ extern const struct wl_interface xdg_wm_base_interface;
  * positioning a surface raises an error.
  */
 extern const struct wl_interface xdg_positioner_interface;
-#endif
-#ifndef XDG_SURFACE_INTERFACE
-#define XDG_SURFACE_INTERFACE
 /**
  * @page page_iface_xdg_surface xdg_surface
  * @section page_iface_xdg_surface_desc Description
@@ -238,9 +230,6 @@ extern const struct wl_interface xdg_positioner_interface;
  * has not been destroyed.
  */
 extern const struct wl_interface xdg_surface_interface;
-#endif
-#ifndef XDG_TOPLEVEL_INTERFACE
-#define XDG_TOPLEVEL_INTERFACE
 /**
  * @page page_iface_xdg_toplevel xdg_toplevel
  * @section page_iface_xdg_toplevel_desc Description
@@ -287,9 +276,6 @@ extern const struct wl_interface xdg_surface_interface;
  * Attaching a null buffer to a toplevel unmaps the surface.
  */
 extern const struct wl_interface xdg_toplevel_interface;
-#endif
-#ifndef XDG_POPUP_INTERFACE
-#define XDG_POPUP_INTERFACE
 /**
  * @page page_iface_xdg_popup xdg_popup
  * @section page_iface_xdg_popup_desc Description
@@ -350,7 +336,6 @@ extern const struct wl_interface xdg_toplevel_interface;
  * for the xdg_popup state to take effect.
  */
 extern const struct wl_interface xdg_popup_interface;
-#endif
 
 #ifndef XDG_WM_BASE_ERROR_ENUM
 #define XDG_WM_BASE_ERROR_ENUM
